move getShowingDetails into separate firestore-methods.ts file
